# perf: added cacheTtl option to match